Specify auto expiration for redis backend
